Hide `Kzg` struct behind `embedded-kzg-settings` feature and invert `subspace-archiving` <-> `subspace-verification` dependency
