+++ crates/subspace-archiving/benches/archiving.rs
@@ -2,7 +2,6 @@ use criterion::{black_box, criterion_group, criterion_main, Criterion};
 use rand::{thread_rng, Rng};
 use std::num::NonZeroUsize;
 use subspace_archiving::archiver::Archiver;
-use subspace_core_primitives::crypto::kzg;
 use subspace_core_primitives::crypto::kzg::Kzg;
 use subspace_core_primitives::pieces::Record;
 use subspace_erasure_coding::ErasureCoding;
@@ -13,7 +12,7 @@ const SMALL_BLOCK_SIZE: usize = 500;
 fn criterion_benchmark(c: &mut Criterion) {
     let mut input = vec![0u8; AMOUNT_OF_DATA];
     thread_rng().fill(input.as_mut_slice());
-    let kzg = Kzg::new(kzg::embedded_kzg_settings());
+    let kzg = Kzg::new();
     let erasure_coding = ErasureCoding::new(
         NonZeroUsize::new(Record::NUM_S_BUCKETS.next_power_of_two().ilog2() as usize)
             .expect("Not zero; qed"),
