+++ crates/subspace-farmer-components/benches/plotting.rs
@@ -4,7 +4,6 @@ use rand::prelude::*;
 use std::env;
 use std::num::{NonZeroU64, NonZeroUsize};
 use subspace_archiving::archiver::Archiver;
-use subspace_core_primitives::crypto::kzg;
 use subspace_core_primitives::crypto::kzg::Kzg;
 use subspace_core_primitives::pieces::Record;
 use subspace_core_primitives::segments::{HistorySize, RecordedHistorySegment};
@@ -30,7 +29,7 @@ fn criterion_benchmark(c: &mut Criterion) {
     let sector_index = 0;
     let mut input = RecordedHistorySegment::new_boxed();
     StdRng::seed_from_u64(42).fill(AsMut::<[u8]>::as_mut(input.as_mut()));
-    let kzg = Kzg::new(kzg::embedded_kzg_settings());
+    let kzg = Kzg::new();
     let erasure_coding = ErasureCoding::new(
         NonZeroUsize::new(Record::NUM_S_BUCKETS.next_power_of_two().ilog2() as usize)
             .expect("Not zero; qed"),
