+++ crates/subspace-farmer-components/benches/proving.rs
@@ -11,7 +11,6 @@ use std::io::Write;
 use std::num::{NonZeroU64, NonZeroUsize};
 use std::{env, fs, slice};
 use subspace_archiving::archiver::Archiver;
-use subspace_core_primitives::crypto::kzg;
 use subspace_core_primitives::crypto::kzg::Kzg;
 use subspace_core_primitives::pieces::Record;
 use subspace_core_primitives::pos::PosSeed;
@@ -57,7 +56,7 @@ pub fn criterion_benchmark(c: &mut Criterion) {
     let mut input = RecordedHistorySegment::new_boxed();
     let mut rng = StdRng::seed_from_u64(42);
     rng.fill(AsMut::<[u8]>::as_mut(input.as_mut()));
-    let kzg = &Kzg::new(kzg::embedded_kzg_settings());
+    let kzg = &Kzg::new();
     let erasure_coding = &ErasureCoding::new(
         NonZeroUsize::new(Record::NUM_S_BUCKETS.next_power_of_two().ilog2() as usize)
             .expect("Not zero; qed"),
