+++ crates/subspace-service/src/sync_from_dsn/piece_validator.rs
@@ -1,12 +1,12 @@
 use async_trait::async_trait;
 use sc_client_api::AuxStore;
 use sc_consensus_subspace::archiver::SegmentHeadersStore;
-use subspace_archiving::archiver::is_piece_valid;
 use subspace_core_primitives::crypto::kzg::Kzg;
 use subspace_core_primitives::pieces::{Piece, PieceIndex};
 use subspace_networking::libp2p::PeerId;
 use subspace_networking::utils::piece_provider::PieceValidator;
 use subspace_networking::Node;
+use subspace_verification::is_piece_valid;
 use tracing::{error, warn};
 
 pub(crate) struct SegmentCommitmentPieceValidator<AS> {
