+++ test/subspace-test-client/src/lib.rs
@@ -34,7 +34,7 @@ use sp_core::{Decode, Encode};
 use std::num::{NonZeroU64, NonZeroUsize};
 use std::slice;
 use std::sync::Arc;
-use subspace_core_primitives::crypto::kzg::{embedded_kzg_settings, Kzg};
+use subspace_core_primitives::crypto::kzg::Kzg;
 use subspace_core_primitives::objects::BlockObjectMapping;
 use subspace_core_primitives::pieces::Record;
 use subspace_core_primitives::pos::PosSeed;
@@ -121,7 +121,7 @@ async fn start_farming<PosTable, Client>(
 {
     let (plotting_result_sender, plotting_result_receiver) = futures::channel::oneshot::channel();
 
-    let kzg = Kzg::new(embedded_kzg_settings());
+    let kzg = Kzg::new();
     let erasure_coding = ErasureCoding::new(
         NonZeroUsize::new(Record::NUM_S_BUCKETS.next_power_of_two().ilog2() as usize)
             .expect("Not zero; qed"),
@@ -205,7 +205,7 @@ where
     PosTable: Table,
     Client: BlockBackend<Block> + HeaderBackend<Block>,
 {
-    let kzg = Kzg::new(embedded_kzg_settings());
+    let kzg = Kzg::new();
     let mut archiver =
         subspace_archiving::archiver::Archiver::new(kzg.clone(), erasure_coding.clone());
